Separate shape doc value format from field implementation
